backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
sqlalchemy
psycopg2-binary
ultralytics
transformers
efficientnet_pytorch
opencv-python
numpy
pillow
python-dotenv
groq
pytest
httpx
jinja2
pdfkit
